feat: 管理者向けユーザーページにユーザー詳細モーダルを追加する
Issue #785 のうち、ユーザー個別情報(処罰状況等)をモーダルから確認・操作できるようにする対応が未実装だったため追加した。権限変更・回答投稿制限の管理は行内操作として分散していたが、詳細モーダルに一本化し、行はユーザー基本情報の表示に専念させた。
